refactor(core): extract tab drag-reorder + animation into TabReorderController
Behaviour-preserving extraction of the ~900-line tab drag-reorder /
smooth-animation subsystem out of the 2,151-line tabs.ts into a dedicated
TabReorderController (operating on Tabs via a narrow ITabReorderHost). Zero
logic changes — validated by the existing suites (tabsAnimation 2,141 lines,
tabs, tabsContainer): 1107 tests / 65 suites green, unchanged.

tabs.ts keeps the DOM event wiring (constructor listeners, per-tab drag/drop
handlers, void/chip handlers) and delegates the reorder logic + `_animState`
to the controller. Isolates the subsystem so the multi-row 2-D cross-row
reorder can be added without tangling it into tabs.ts. No public API change;
the `dv-tabs-container--wrap` reorder guard is preserved.
